Fixed Earth and Space Icon
The if/else statement determining which icon to use tested first for science (and thus displayed a beaker) before testing for earth and space. The order has been changed in order to test for the string 'earth' or 'space' before testing for 'science', thus now showing the star icon that is used for space.