Fixed calendar so that it works for 2016-2017
Colonel By has migrated to a new Google Calendar for the 2016-2017 school year, and this app has been updated in order to use that as the source. Unfortunately, it seems to only have Day 1's on there at the moment; I may have to make my own calendar to restore full functionality.